georectification

Documentation for georectification.

georectification.Coords!

georectification.HSI_2_RGB

georectification.Rotate

georectification.Spec_2_RGB

georectification.generateDerivedMetrics!

georectification.generateReflectance!

georectification.georectify

georectification.getBackgroundTile

georectification.getHSIFlightData

georectification.getHdrFile

georectification.getIMUdata

georectification.getTimes

georectification.ilat_ilon!

georectification.ilat_ilon!

georectification.imagecoords

georectification.imagecoordsFlipped

georectification.masterLCF

georectification.plot_background

georectification.readToDataFrame

georectification.Coords! — Method

Coords!(df::DataFrame,
        lcf_path::String,
        z_ground::Float64,
        θ::Float64,
        flipped::Bool)

Compute the ground coordinates for each pixel of an HSI df. Update the original df to include

latitude

longitude

altitude

pixeltimes

roll

pitch

heading

georectification.HSI_2_RGB — Method

HSI_2_RGB(λs_spec::AbstractVector, HSI::AbstractMatrix, d::Int)

Convert reflectance data from an HSI to an RGB representation.

Arguments

λs_spec::AbstractVector`: Wavelength values in nm of the spectrum

HSI::AbstractMatrix: Matrix of reflectance or radiance values with shape npixels x nbands

d::Int: 50, 55, 65, 75. Determines the illuminant used. If in doubt, use d=65.

Source: M. Magnusson, J. Sigurdsson, S. E. Armansson, M. O. Ulfarsson, H. Deborah and J. R. Sveinsson, "Creating RGB Images from Hyperspectral Images Using a Color Matching Function," IGARSS 2020 - 2020 IEEE International Geoscience and Remote Sensing Symposium, 2020, pp. 2045-2048, doi: 10.1109/IGARSS39084.2020.9323397.

georectification.Rotate — Method

Rotate(heading, pitch, roll)

Return a rotation matrix corresponding to the given orientation angles.

georectification.Spec_2_RGB — Method

Spec_2_RGB(λs_spec::AbstractVector, Spectrum::AbstractVector, d::Int)

Convert a pixel spectrum from an HSI to an RGB representation.

Arguments

λs_spec::AbstractVector`: Wavelength values in nm of the spectrum

Spectrum::AbstractVector: Vector of reflectance or radiance values

d::Int: 50, 55, 65, 75. Determines the illuminant used. If in doubt, use d=65.

Source: M. Magnusson, J. Sigurdsson, S. E. Armansson, M. O. Ulfarsson, H. Deborah and J. R. Sveinsson, "Creating RGB Images from Hyperspectral Images Using a Color Matching Function," IGARSS 2020 - 2020 IEEE International Geoscience and Remote Sensing Symposium, 2020, pp. 2045-2048, doi: 10.1109/IGARSS39084.2020.9323397.

georectification.generateDerivedMetrics! — Method

generateDerivedMetrics!(df::DataFrame, λs::Array{Float64})

Given a dataframe df with HSI data, compute a variety of derived wavelength indices such as the NDVI.

georectification.generateReflectance! — Method

generateReflectance!(df::DataFrame, specPath::String, specHdrPath::String, calibrationPath::String, wavelengths::Array{Float64, 1})

Given a dataframe df containing the radiance data, generate reflectance data using collected irradiance spectrum and Resonon-supplied calibration data.

NOTE

The conversion assumes a Lambertian surface. This is clearly not the case for water that isn't perfectly still.

georectification.georectify — Method

georectify(bilpath::String,
           bilhdrpath::String,
           timespath::String,
           specpath::String,
           spechdrpath::String,
           calibrationpath::String,
           lcfpath::String,
           λs::Array{Float64},
           z_ground::Float64,
           θ_view::Float64,
           isFlipped::Bool,
           ndigits::Int,
           )

Generate a georectified dataframe including radiance, reflectance, derived metrics, and position data.

georectification.getBackgroundTile — Method

getBackgroundTile(w::Float64, n::Float64, e::Float64, s::Float64, outName::String)

Grab Esri World Imagery tiles for region with corners (w,n), (e,s) in longitude and latitude. Saves resulting image to outpath

Note: result images are saved in Web-Mercator projection by default. See WebMercatorfromLLA and LLAfromWebMercator from Geodesy.jl for conversion details.

georectification.getHSIFlightData — Method

getHSIFlightData(df::DataFrame, lcf_path::String)

Given IMU data found at lcf_path, compute the position and orientation of the Aerial Vehicle at time of capture for each pixel.

georectification.getHdrFile — Method

getHdrFile(pathToHdrFile::String)

Read an ENVI-style header file and return a dictionary with image metadata.

georectification.getIMUdata — Method

getIMUdata(pathToLCF::String)

Read a .lcf file and return a DataFrame df with corresponding IMU data and the value start_time indicating the start time of the clock in linux epoch time.

Data Returned

time: system time in seconds starting at zero

roll: angle in radians.

pitch: angle in radians.

heading: aka yaw. angle in radians.

longitude: GPS longitude in degrees, negative for west longitudes

latitude: GPS latitude in degrees

altitude: height in meters above the WGS-54 ellipsoid

x: x coordinate in meters in UTMz coordinate system

y: y coordinate in meters in UTMz coordinate system

z: z coordinate in meters in UTMz coordinate system

zone: the UTM zone for the local coordinate system

isnorth: boolean for whether or not position is above equator

NOTES

The conversion between GPS time and Unix Epoch time was a little confusing. It is good to double check that this is correct (how should we deal with leap seconds?). This is relevant for parsing data from the FLIR as well.

georectification.getTimes — Method

getTimes(pathToTimes::String)

Read a .times and return values starting at t=0. Times are assumed to be aligned with the corresponding .lcf file.

georectification.ilat_ilon! — Method

ilat_ilon!(df::DataFrame, ndigits)

Update a dataframe to include the rounded coordinates ilat and ilon with ndigits of precision.

georectification.ilat_ilon! — Method

ilat_ilon!(df::DataFrame)

Dispatch of ilat_ilon() to use default value of 6 for ndigits.

georectification.imagecoords — Method

imagecoords(i, j, N, f)

Given pixel indices i and j, compute the coordinates of a HSI pixel in the image coordinate system. The height is the focal length f, and the pixels are assumed to lie along the y-axis.

georectification.imagecoordsFlipped — Method

imagecoordsFlipped(i, j, N, f)

Same as imagecoord() but with the y axis flipped. This is useful in case the camera settings for direction of flight are backwards.

georectification.masterLCF — Method

masterLCF(folder::String)

Loop through all .lcf files in folder and return a DataFrame containing the start and end times for each datacube aquisition.

Fields

paths: Paths to each HSI directory

files: file name for .lcf files

tstart: starting time for each datacube acquisition

tend: ending time for each datacube acquisition

georectification.plot_background — Method

plot_background(sat_tile, longitudes, latitudes)

Given a satellite background tile sat_tile and associated longitude and latitude ranges, construct a plot

georectification.readToDataFrame — Function

readToDataFrame(pathToBinaryFile::String, pathToHdrFile::String, pathToTimesFile::String)

Read ENVI file with corresponding header (.hdr) and times (.times) file and return a DataFrame containing the image data flattened to (nbands, ncols*nrows).